refactor: adding checks for listing/id/parameter PUT endpoint

+++ apps/marketplace/src/pages/api/v1/listings/[id]/[parameters]/index.ts
@@ -1,7 +1,7 @@
 import { NextApiRequest, NextApiResponse } from 'next';
 import { apiHandler, formatAPIResponse } from '@/utils/api';
 import PrismaClient, { ListingsParametersValue } from '@inc/db';
-import { NotFoundError } from '@/errors';
+import { ForbiddenError } from '@/errors/AuthError';
 import * as z from 'zod';
 import { Parameter } from '@inc/db';
 import { parseListingId } from '@/utils/api';
@@ -24,11 +24,7 @@ function formatParametersResponse(parameters: Parameter[]): any[] {
  * @param id The listing id
  * @returns The listing if it exists
  */
-
-
 const getListingParameters = async (req: NextApiRequest, res: NextApiResponse) => {
-
-
     const id = parseListingId(req.query.id as string);
     const listing = await checkListingExists(id);
 
@@ -55,38 +51,53 @@ const parameterSchema = z.object({
 const requestSchema = z.object({
     parameters: z.array(parameterSchema),
 });
-
-const updateListingParameters = async (req: NextApiRequest, res: NextApiResponse) => {
-    const id = parseListingId(req.query.id as string);
-    const listing = await checkListingExists(id);
-
-    // Assuming that the request body contains the new parameters
-    const { parameters } = requestSchema.parse(req.body);
-
-    // Delete all the existing parameters for this listing
-    await PrismaClient.listingsParametersValue.deleteMany({
-        where: {
-            listingId: id,
-        },
-    });
-
-    // Add the new parameters to the database
-    await Promise.all(parameters.map(async (param: ListingsParametersValue) => {
-        await PrismaClient.listingsParametersValue.create({
-            data: {
-                listingId: id,
-                parameterId: param.parameterId,
-                value: param.value,
-            },
-        });
-    }));
-    res.status(200).json(formatAPIResponse({ success: true, data: 'Listing parameters updated successfully' }));
-};
-
-
-
-
+interface NextApiRequestWithToken extends NextApiRequest {
+    token?: {
+        user: {
+            id: string;
+            role: number;
+            company: string;
+        };
+    };
+}
 export default apiHandler()
     .get(getListingParameters)
-    .put(updateListingParameters);
+    .put(async (req, res) => {
+        const id = parseListingId(req.query.id as string);
+        const userId = req.token?.user?.id;
+        const userRole = req.token?.user?.role;
+
+        const listing = await checkListingExists(id);
+
+        const isOwner = listing.owner === userId;
+        const isAdmin = userRole && userRole >= 1;
+        const sameCompany = req.token?.user?.company === listing.users.companyId;
+
+        if (!isOwner && !isAdmin && !sameCompany) {
+            throw new ForbiddenError();
+        }
+
+        // Assuming that the request body contains the new parameters
+        const { parameters } = requestSchema.parse(req.body);
+
+        // Delete all the existing parameters for this listing
+        await PrismaClient.listingsParametersValue.deleteMany({
+            where: {
+                listingId: id,
+            },
+        });
+
+        // Add the new parameters to the database
+        await Promise.all(parameters.map(async (param: ListingsParametersValue) => {
+            await PrismaClient.listingsParametersValue.create({
+                data: {
+                    listingId: id,
+                    parameterId: param.parameterId,
+                    value: param.value,
+                },
+            });
+        }));
+
+        res.status(200).json(formatAPIResponse({ success: true, data: 'Listing parameters updated successfully' }));
+    };);
